requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==15.0.1
pydantic==2.11.4
aiosqlite==0.21.0
httpx==0.28.1
python-dotenv==1.1.0
pywebview>=6.1,<7
pystray==0.19.5
Pillow==10.4.0
python-multipart==0.0.9
